requirements copy.txt
# numpy==1.19
# # numpy==1.24.*
# # mmcv==1.5.3
# mmcv-full==1.5.3
# # mmcv==1.6
# # mmcv-full==1.6.0
# mmdet==2.25.0
# # mmpose==0.27.0
# mmpose==0.29
# pip-chill==1.0.1
# opencv-contrib-python==4.4.*

numpy
# numpy==1.24.*
# mmcv==1.5.3
mmcv-full
# mmcv==1.6
# mmcv-full==1.6.0
# mmdet==2.25.0
# mmpose==0.27.0
mmpose==0.29
# pip-chill==1.0.1
opencv-python


# numpy==1.*
# mmcv==1.*
# mmcv-full==1.*
# mmdet==2.*
# mmpose==0.*
# pip-chill==1.0.1
# opencv-contrib-python>=4.5.5